Implement the store router's update route
A seller should be able to update their store. Therefore, a /store/update route
has been created to solve this same issue. This also led to the creation of the store.js router.
Authentication of the user also included as middlewares to verify that a user is a seller and has access to the specific store he or she is attempting to update.