implement product and store tags
- Uncommented the Tag.js model
- added a tag property to the Store.js and Product.js models

This is to make sure that a store and a product can contain a tag
It makes it easier for a buyer to notice any special features that the
store or product possess if it is not provided in the available
categories. It also allows for a more seller defined organization of
products apart from the standard categories.