Create feature for seller to add tags to a product on creation
A product should be able to have many tags to specifiy special features about it.
A seller should, therefore, be able to add many tags to a product when creating one.
tags shouldn't be duplicated whenn created and saved in the database, therefore, a check is necessary to make sure no duplicate tags are saved
to the databse.