Implement hot reloading of browser
To improve the quality and speed of development,
hot reloading or browser when server files change has to be implemented.
It is implemented with the help of the npm fast-live-reload package.

An flr script was added to package.json to enable easy start.
This functionality or package is implemented to work alongside
nodemone, and laravel-mix in a node-react app.

fast-live-reload also doent require any changes to be made to the existing app code
and this is super convinient as no code has to be altered when pushing for production.
It is also the only package so far with live reloading that hasnt proved buggy with the app.