Add support for local fonts
Added support for local fonts to imporve local offline development.
This is also to increase the chances of the fonts loading
even if the cloud service might not be available. In short
it increases the availability of the fonts to our platform.

Also make some changes to the sign up for UI