Finished designing the buyer sign up ui
A buyer should be able to signup with as little information as possible.
This is to prevent the user from aborting the signup operation
due to unwaranted difficulties or loss of interest.
A buyer will be able to sign up and join the gulait platform.
Allowing them to save and track their interactions on the platform for their convinience.
The platform will also be able to know the consistent users on the platform and
figure out their reach in the world.