Hide the input form text elements
Hide the input form text elements and only show
them when the input element has been focused on.
This is to keep the form short and as neat as possible
while still providing the necessary information when needed.
Avoided using popovers because they dont scale down well on mobile screens